222: Automatic share distribution
